moved data-model tests
